+++ services/web/client/src/components/character/tabContents/general/form/NpcSkillsEdit.tsx
@@ -53,9 +53,8 @@ export default function NpcSkillsEdit({ form, accentColor }: NpcSkillsEditProps)
   ] as const;
 
   return (
-    <div className="grid grid-cols-1 sm:grid-cols-2 gap-2">
+    <div className="grid grid-cols-1 gap-2">
       {skills.map(({ key, icon: Icon, ability }) => {
-        const skillValue = form.watch(`stats.skills.${key}`) || 0;
         const abilityScore = form.watch(`stats.abilityScores.${ability}`) || 10;
         const abilityModifier = Math.floor((abilityScore - 10) / 2);
