+++ services/web/client/src/components/character/tabContents/general/view/PlayerGeneralTabContent.tsx
@@ -6,9 +6,14 @@ import AbilityScores from "@/components/character/tabContents/general/shared/Abi
 import { Checkbox } from "@/components/ui/checkbox";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
 import AbilitiesSection from "@/components/character/tabContents/shared/AbilitiesSection";
-import SavingThrow from "@/components/character/tabContents/general/shared/SavingThrow";
-import Skills from "../shared/Skills";
+import { Bird, Mountain, Shovel, Waves } from "lucide-react";
 import Column2 from "./Column2";
+import CharacterHealthBar from "@/components/character/CharacterHealthBar";
+import ShieldIcon from "@public/assets/icons/shield-icon.svg";
+import FeatherIcon from "@public/assets/icons/feather-icon.svg";
+import RunningIcon from "@public/assets/icons/running-icon.svg";
+import Image from "next/image";
+import Statistics from "../../shared/Statistics";
 
 interface PlayerGeneralTabContentProps {
   player: Player;
@@ -17,12 +22,77 @@ interface PlayerGeneralTabContentProps {
 
 export default function PlayerGeneralTabContent({ player, accentColor }: PlayerGeneralTabContentProps) {
   const t = useTranslations("characterDetail.player.general");
+  const tbattle = useTranslations("characterDetail.battle");
   const tPlayer = useTranslations("characterDetail.player");
   const tAlignment = useTranslations("alignments");
   const tClass = useTranslations("classes");
   const tEdit = useTranslations("characterDetail.edit");
 
   const [checked, setChecked] = useState<boolean>(player.inspiration);
+
+  const speedBadges = [
+    {
+      key: "walk",
+      value: player.stats.speed.walk,
+      icon: (
+        <Image
+          src={RunningIcon}
+          alt=""
+          aria-hidden="true"
+          className="size-6"
+        />
+      ),
+      tooltipKey: "walkSpeedTooltip",
+    },
+    {
+      key: "climb",
+      value: player.stats.speed.climb,
+      icon: (
+        <Mountain
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "climbSpeedTooltip",
+    },
+    {
+      key: "swim",
+      value: player.stats.speed.swim,
+      icon: (
+        <Waves
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "swimSpeedTooltip",
+    },
+    {
+      key: "fly",
+      value: player.stats.speed.fly,
+      icon: (
+        <Bird
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "flySpeedTooltip",
+    },
+    {
+      key: "burrow",
+      value: player.stats.speed.burrow,
+      icon: (
+        <Shovel
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "burrowSpeedTooltip",
+    },
+  ];
 
   function infoExhaustionLevel(level: number): string {
     return t(`exhaustionLevels.${level}`);
@@ -95,6 +165,12 @@ export default function PlayerGeneralTabContent({ player, accentColor }: PlayerG
               ))}
             </dl>
           </Card>
+
+          {/* Statistiques */}
+          <Statistics
+            player={player}
+            accentColor={accentColor}
+          />
 
           <Column2
             player={player}
