+++ services/web/client/src/components/character/tabContents/shared/NpcStatistics.tsx
@@ -1,0 +1,169 @@
+import { Card } from "@/components/ui/card";
+import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
+import Image from "next/image";
+import { Bird, Mountain, Shovel, Waves } from "lucide-react";
+import ShieldIcon from "@public/assets/icons/shield-icon.svg";
+import RunningIcon from "@public/assets/icons/running-icon.svg";
+import FeatherIcon from "@public/assets/icons/feather-icon.svg";
+import { NPC } from "@/types/character";
+import { useTranslations } from "next-intl";
+import CharacterHealthBar from "../../CharacterHealthBar";
+
+interface NpcStatisticsProps {
+  npc: NPC;
+  accentColor: string;
+}
+
+export default function NpcStatistics({ npc, accentColor }: NpcStatisticsProps) {
+  const t = useTranslations("characterDetail.battle");
+
+  const speedBadges = [
+    {
+      key: "walk",
+      value: npc.stats.speed.walk,
+      icon: (
+        <Image
+          src={RunningIcon}
+          alt=""
+          aria-hidden="true"
+          className="size-6"
+        />
+      ),
+      tooltipKey: "walkSpeedTooltip",
+    },
+    {
+      key: "climb",
+      value: npc.stats.speed.climb,
+      icon: (
+        <Mountain
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "climbSpeedTooltip",
+    },
+    {
+      key: "swim",
+      value: npc.stats.speed.swim,
+      icon: (
+        <Waves
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "swimSpeedTooltip",
+    },
+    {
+      key: "fly",
+      value: npc.stats.speed.fly,
+      icon: (
+        <Bird
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "flySpeedTooltip",
+    },
+    {
+      key: "burrow",
+      value: npc.stats.speed.burrow,
+      icon: (
+        <Shovel
+          size={24}
+          className="text-black"
+          aria-hidden="true"
+        />
+      ),
+      tooltipKey: "burrowSpeedTooltip",
+    },
+  ];
+  return (
+    <Card
+      className="gap-3 p-4 md:px-6 col-span-3 2xl:col-span-1 h-fit"
+      role="region"
+      aria-labelledby="stats-heading-npc">
+      <h2
+        id="stats-heading-npc"
+        className={`text-xl sm:text-2xl font-semibold ${accentColor}`}>
+        {t("stats")}
+      </h2>
+      <div className="flex flex-row justify-start gap-2 text-xl font-extrabold flex-wrap">
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <div
+              className="bg-white text-black flex flex-row justify-center gap-1 rounded-full p-2 items-center cursor-help focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-2"
+              tabIndex={0}
+              role="img"
+              aria-label={`${t("armorClass")} ${npc.stats.armorClass}`}>
+              <Image
+                src={ShieldIcon}
+                alt=""
+                aria-hidden="true"
+                width={20}
+                height={20}
+                className="size-5"
+              />
+              <span aria-hidden="true">{npc.stats.armorClass}</span>
+            </div>
+          </TooltipTrigger>
+          <TooltipContent>{t("armorClassTooltip")}</TooltipContent>
+        </Tooltip>
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <div
+              className="bg-white text-black flex flex-row justify-center gap-1 rounded-full p-2 items-center cursor-help focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-2"
+              tabIndex={0}
+              role="img"
+              aria-label={`${t("initiativeTooltip")} ${npc.stats.initiative > 0 ? `+${npc.stats.initiative}` : npc.stats.initiative}`}>
+              <Image
+                src={FeatherIcon}
+                alt=""
+                aria-hidden="true"
+                width={20}
+                height={20}
+                className="size-5"
+              />
+              <span aria-hidden="true">
+                {npc.stats.initiative > 0 ? `+${npc.stats.initiative}` : npc.stats.initiative}
+              </span>
+            </div>
+          </TooltipTrigger>
+          <TooltipContent>{t("initiativeTooltip")}</TooltipContent>
+        </Tooltip>
+        {speedBadges.map(
+          (badge) =>
+            badge.value > 0 && (
+              <Tooltip key={badge.key}>
+                <TooltipTrigger asChild>
+                  <div
+                    className="bg-white text-black flex flex-row justify-center gap-1 rounded-full p-2 items-center cursor-help focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-2"
+                    tabIndex={0}
+                    role="img"
+                    aria-label={`${t(badge.tooltipKey as any)} ${badge.value} ${t("feet")}`}>
+                    {badge.icon}
+                    <span aria-hidden="true">
+                      {badge.value}
+                      {t("feetAbbr")}
+                    </span>
+                  </div>
+                </TooltipTrigger>
+                <TooltipContent>{t(badge.tooltipKey as any)}</TooltipContent>
+              </Tooltip>
+            ),
+        )}
+      </div>
+      <CharacterHealthBar
+        currentHP={npc.stats.currentHitPoints}
+        maxHP={npc.stats.maxHitPoints}
+        tempHP={npc.stats.tempHitPoints}
+      />
+      <div className="text-lg px-2">
+        <span>{t("hitPointsRoll")}</span>
+        {npc.hitPointsRoll && <span className="font-bold"> {npc.hitPointsRoll}</span>}
+      </div>
+    </Card>
+  );
+}
